+++ crates/pallets/starknet/src/lib.rs
@@ -1190,4 +1190,7 @@ impl<T: Config> Pallet<T> {
     pub fn chain_id() -> Felt252Wrapper {
         T::ChainId::get()
     }
+    pub fn is_transaction_fee_disabled() -> bool {
+        T::DisableTransactionFee::get()
+    }
 }